+++ tests/unit/backends/test_openflow13.py
@@ -1,6 +1,6 @@
 """ Tests for /backends/openflow13.py """
 
-from unittest.mock import MagicMock, patch
+from unittest.mock import MagicMock, patch, AsyncMock
 
 from napps.amlight.sdntrace import settings
 from napps.amlight.sdntrace.backends.openflow13 import (
@@ -52,22 +52,22 @@ class TestOpenflow13:
         assert in_port == 0
         assert switch == 0
 
-    def test_packet_out_with_color(self):
+    async def test_packet_out_with_color(self):
         """Test packet in with color ee:ee:ee:ee:ee:01."""
         data = b"\x00\x15\xaf\xd58\x98\xee\xee\xee\xee\xee\x01\x08\x00testdata"
         # Ethernet(destination='00:15:af:d5:38:98',
         #          source='ee:ee:ee:ee:ee:01', ether_type=0x800,
         #          data=b'testdata')
-        controller = MagicMock()
+        controller = AsyncMock()
         switch = MagicMock()
         switch.connection.value = "00:15:af:d5:38:98"
         port = 1
 
-        send_packet_out(controller, switch, port, data)
+        await send_packet_out(controller, switch, port, data)
 
         # Verify that the controller sent the packet_out
-        controller.buffers.msg_out.put.assert_called_once()
-        called_event = controller.buffers.msg_out.put.call_args.args[0]
+        controller.buffers.msg_out.aput.assert_called_once()
+        called_event = controller.buffers.msg_out.aput.call_args.args[0]
 
         # Verify the packet_out values
         assert called_event.name == "kytos/of_lldp.messages.out.ofpt_packet_out"
@@ -77,9 +77,9 @@ class TestOpenflow13:
         assert called_event.content["message"].actions[0].port == settings.OFPP_TABLE_13
 
     @patch("napps.amlight.sdntrace.backends.openflow13.of_msg_prio")
-    def test_send_packet_out(self, mock_of_msg_prio) -> None:
+    async def test_send_packet_out(self, mock_of_msg_prio) -> None:
         """Test send_packet_out."""
-        controller = MagicMock()
-        send_packet_out(controller, MagicMock(), MagicMock(), MagicMock())
-        assert controller.buffers.msg_out.put.call_count == 1
+        controller = AsyncMock()
+        await send_packet_out(controller, MagicMock(), MagicMock(), MagicMock())
+        assert controller.buffers.msg_out.aput.call_count == 1
         assert mock_of_msg_prio.call_count == 1